Read Cds table with header taken from a ReadMe file.
When ``Cds`` reader object is created with a
``readme`` parameter passed to it at initialization, then when
the ``read`` method is executed with a table filename, the header
information for the specified table is taken from the ``readme`` file.
An ``IncosistentTableError`` is raised if the ``readme`` file does not
have header information for the given table.

    >>> readme = "t/vizier/ReadMe"
    >>> r = asciitable.Cds(readme)
    >>> table = r.read("t/vizier/table1.dat")

At initialization, ``CdsHeader`` stores name of "ReadMe" file in
``self.readme``. This initialization takes place when the ``Cds``
class is initialized to get a new reader.

When ``read`` is invoked ``BaseReader`` stores name of table filename
in ``self.data.table_name``; only if ``table`` is a file.

Within ``CdsData`` the ``get_data_lines`` function checks to see if
``self.table_name`` and ``self.header.readme`` are true and if so,
the step of skipping over the header information in ``lines``
is skipped. The ``lines`` attribute will not have any header information.

Within ``CdsHeader`` the ``get_cols`` method checks to see if both
``self.readme`` and ``self.data.filename`` are true. If they are, then
the ``lines`` parameter is ignored and a new ``lines`` list is
read from ``self.readme`` and will contain only the header of the
``self.data.table_name`` table. Parsing of this set of lines takes place
as usual.

Since ``self.data.table_name`` and ``self.header.readme`` of the ``Cds``
reader are set to None by default, if not ``readme`` is specified at
initialization, then the regular parsing of the table takes place.
Therefore any previous code that uses ``Cds`` reader will continue to
work.