Add machinery to use data types and null vals specified within table
Several supported table formats include specification of the
column data types.  Previously, asciitable was ignoring those
and always doing automatic type detection based on table values.
Now for RDB, CDS, and IPAC tables the column types within
the table are used.  Other formats (notably DAOPHOT) could be
extended similarly.

This version includes an API change that is not backward compatible
but unlikely to affect anyone.  The Outputter classes (BaseOutputter
and NumpyOutputter) both define a class variable "default_converter"
to specify the default conversion routines.  This was previously
a single list containing conversion routines to try in order.  Now
it is a dictionary with keys 'any', 'num', 'float', 'int', 'str',
where each value is a list of conversion routines to try in order.

Null values are now fully supported for CDS and IPAC tables.