add fill_values etc. for writer objects
asciitable.write now accepts
- fill_values
- fill_include_names
- fill_exclude_names
The format is identical to same fuctionality for Readers.

To make this easier, column objects are now iterable (just like lists or
numpy arrays). They return their values formatted with self.formatter.
This can be used in the future to unify converters (for reading) and
formaters (for writing).