Changed default BaseReader.inconsistent_handler to return str_vals instead of [].
Nosetests after the inconsistent_handler patch showed a failure for
reading CDS files (t/cds.dat).  This was connected to
inconsistent_handler changing str_vals to [] for a row with
inconsitent columns.  Changed the routine so that the default is to
pass back the original str_vals so that it has the intended "do-nothing"
behavior.