Fix problems in column type management.
The Memory class was not using the right set of default_converters so that
types determined in the read process were not propagated through the write
process.  In the has_numpy == False case there was a bug in the way types
determined so it was always IntType.

Added several new tests of type functionality and fixed a mistake in the
expected output for an Rdb write test.

** There is still an open problem that is identified through a fail in nosestests.