Add n_data_cols property to BaseHeader
Previously BaseReader was using len(header.names) as the
for the number of data columns to expect from split.  This
doesn't work for FixedWidth tables where the splitter only
returns column data for the output columns.  The property
now allows derived classes to set this explicitly if it
will be different from len(self.names).